chore: move predict_clr to celery task
