+++ app/grants/management/commands/estimate_clr.py
@@ -23,8 +23,8 @@ from django.core.management.base import BaseCommand
 from django.utils import timezone
 
 from dashboard.utils import get_tx_status, has_tx_mined
-from grants.clr import predict_clr
 from grants.models import Contribution, Grant, GrantCLR
+from grants.tasks import predict_clr
 from marketing.mails import warn_subscription_failed
