+++ app/grants/tasks.py
@@ -10,8 +10,11 @@ from app.services import RedisService
 from celery import app, group
 from celery.utils.log import get_task_logger
 from dashboard.models import Profile
+from grants.clr import calculate_clr_for_donation, fetch_data, populate_data_for_clr
 from grants.models import Grant, Subscription
 from marketing.mails import new_grant, new_grant_admin, new_supporter, thank_you_for_supporting
+from marketing.models import Stat
+from perftools.models import JSONStore
 from townsquare.models import Comment
 
 logger = get_task_logger(__name__)
@@ -19,6 +22,7 @@ logger = get_task_logger(__name__)
 redis = RedisService().redis
 
 CLR_START_DATE = dt.datetime(2020, 9, 14, 15, 0) # TODO:SELF-SERVICE
+
 
 @app.shared_task(bind=True, max_retries=1)
 def update_grant_metadata(self, grant_id, retry: bool = True) -> None:
@@ -50,6 +54,7 @@ def update_grant_metadata(self, grant_id, retry: bool = True) -> None:
                 instance.monthly_amount_subscribed += subscription.get_converted_monthly_amount()
 
     from django.contrib.contenttypes.models import ContentType
+
     from search.models import SearchResult
     if instance.pk:
         SearchResult.objects.update_or_create(
@@ -191,11 +196,13 @@ def process_grant_contribution(self, grant_id, grant_slug, profile_id, package, 
 
         update_grant_metadata.delay(grant_id)
 
+
 @app.shared_task(bind=True, max_retries=1)
 def recalc_clr(self, grant_id, retry: bool = True) -> None:
     obj = Grant.objects.get(pk=grant_id)
-    from grants.clr import predict_clr
     from django.utils import timezone
+
+    from grants.clr import predict_clr
     for clr_round in obj.in_active_clrs.all():
         network = 'mainnet'
         predict_clr(
@@ -205,6 +212,116 @@ def recalc_clr(self, grant_id, retry: bool = True) -> None:
             network=network,
             only_grant_pk=obj.pk
         )
+
+
+@app.shared_task(bind=True, max_retries=1)
+def predict_clr(save_to_db=False, from_date=None, clr_round=None, network='mainnet', only_grant_pk=None):
+    # setup
+    clr_calc_start_time = timezone.now()
+    debug_output = []
+
+    # one-time data call
+    total_pot = float(clr_round.total_pot)
+    v_threshold = float(clr_round.verified_threshold)
+    uv_threshold = float(clr_round.unverified_threshold)
+
+    grants, contributions = fetch_data(clr_round, network)
+
+    if contributions.count() == 0:
+        print(f'No Contributions for CLR {clr_round.round_num}. Exiting')
+        return
+
+    grant_contributions_curr = populate_data_for_clr(grants, contributions, clr_round)
+
+    if only_grant_pk:
+        grants = grants.filter(pk=only_grant_pk)
+
+    # calculate clr given additional donations
+    for grant in grants:
+        # five potential additional donations plus the base case of 0
+        potential_donations = [0, 1, 10, 100, 1000, 10000]
+        potential_clr = []
+
+        for amount in potential_donations:
+            # calculate clr with each additional donation and save to grants model
+            # print(f'using {total_pot_close}')
+            predicted_clr, grants_clr, _, _ = calculate_clr_for_donation(
+                grant,
+                amount,
+                grant_contributions_curr,
+                total_pot,
+                v_threshold,
+                uv_threshold
+            )
+            potential_clr.append(predicted_clr)
+
+        if save_to_db:
+            _grant = Grant.objects.get(pk=grant.pk)
+            clr_prediction_curve = list(zip(potential_donations, potential_clr))
+            base = clr_prediction_curve[0][1]
+            _grant.last_clr_calc_date = timezone.now()
+            _grant.next_clr_calc_date = timezone.now() + timezone.timedelta(minutes=20)
+
+            can_estimate = True if base or clr_prediction_curve[1][1] or clr_prediction_curve[2][1] or clr_prediction_curve[3][1] else False
+
+            if can_estimate :
+                clr_prediction_curve  = [[ele[0], ele[1], ele[1] - base] for ele in clr_prediction_curve ]
+            else:
+                clr_prediction_curve = [[0.0, 0.0, 0.0] for x in range(0, 6)]
+
+            JSONStore.objects.create(
+                created_on=from_date,
+                view='clr_contribution',
+                key=f'{grant.id}',
+                data=clr_prediction_curve,
+            )
+            clr_round.record_clr_prediction_curve(_grant, clr_prediction_curve)
+            
+            try:
+                if clr_prediction_curve[0][1]:
+                    Stat.objects.create(
+                        created_on=from_date,
+                        key=_grant.title[0:43] + "_match",
+                        val=clr_prediction_curve[0][1],
+                        )
+                    max_twitter_followers = max(_grant.twitter_handle_1_follower_count, _grant.twitter_handle_2_follower_count)
+                    if max_twitter_followers:
+                        Stat.objects.create(
+                            created_on=from_date,
+                            key=_grant.title[0:43] + "_admt1",
+                            val=int(100 * clr_prediction_curve[0][1]/max_twitter_followers),
+                            )
+
+                if _grant.positive_round_contributor_count:
+                    Stat.objects.create(
+                        created_on=from_date,
+                        key=_grant.title[0:43] + "_pctrbs",
+                        val=_grant.positive_round_contributor_count,
+                        )
+                if _grant.amount_received_in_round:
+                    Stat.objects.create(
+                        created_on=from_date,
+                        key=_grant.title[0:43] + "_amt",
+                        val=_grant.amount_received_in_round,
+                        )
+            except:
+                pass
+
+            if from_date > (clr_calc_start_time - timezone.timedelta(hours=1)):
+                _grant.save()
+
+        debug_output.append({'grant': grant.id, "clr_prediction_curve": (potential_donations, potential_clr), "grants_clr": grants_clr})
+
+    try :
+        Stat.objects.create(
+            key= clr_type + '_grants_round_6_saturation',
+            val=int(CLR_PERCENTAGE_DISTRIBUTED),
+        )
+    except:
+        pass
+
+    return debug_output
+
 
 @app.shared_task(bind=True, max_retries=3)
 def process_grant_creation_email(self, grant_id, profile_id):
